# Various performance tweaks