--- a/src/brsTypes/Callable.ts
+++ b/src/brsTypes/Callable.ts
@@ -6,6 +6,7 @@ import { Identifier, Token } from '../lexer';
 import { SourceNode } from 'source-map';
 import { Range } from 'vscode-languageserver';
 import { TranspileState } from '../parser/TranspileState';
+import util from '../util';
 
 /** An argument to a BrightScript `function` or `sub`. */
 export interface Argument {
@@ -55,7 +56,7 @@ export class StdlibArgument implements Argument {
     }
 
     /** A fake location exists only within the BRS runtime. */
-    static InternalRange = Range.create(-1, -1, -1, -1);
+    static InternalRange = util.createRange(-1, -1, -1, -1);
 }
 
 export class FunctionParameter {