`vscode-languageserver`'s `Range.create()` seems to be significantly slower than creating the ranges manually. 
![image](https://user-images.githubusercontent.com/2544493/94374228-be441180-00d8-11eb-8c38-da1c4e7b42a0.png)

So, by switching the Lexer to use our own `createRange` function, we can improve its performance. With this change, scanning a 400 line file went from 746 ops/sec to 931 ops/sec
![image](https://user-images.githubusercontent.com/2544493/94374490-95248080-00da-11eb-8a6f-ad0c837100d0.png)

All benchmarks were performed against a 400 line brighterscript file. 

**Notable Changes:**

1. eliminate `Range.create()` in favor of `util.createRange()` and `util.createRangeFromPositions()`. 
    - ~25% increase in `Lexer` performance
    - ~12% increase in `Parser` performance
2. In `Lexer.ts`, don't return previous character in the `advance()` method.
    - ~3% increase in `Lexer` performance
3. In `Lexer.ts`, replace switch statement with a static map.
    - ~27% increase in `Lexer` performance


Overall, this has improved the performance of the the `Lexer+Parser` cycle by ~16% (from 300 ops/sec to 350 ops/sec) for my 400 line test file.