@TwitchBronBron 2020-09-28
I'm sure there are more improvements we can make, bit these are a good start.